// src/modules/merchandising/reports/report/emblishment-status-report/emblishment-status-report-type.ts
export interface IEmblishmentStatusReport {
  COMPANY_NAME: string;
  PONO: string;
  BUYER_NAME: string;
  STYLENO: string;
  PO_QTY: number;
  SECTION_NAME: string;
  WORKORDERQTY: number;
  WO_NO: string;
  SENDQTY: number;
  RECEIVEQTY: number;
  COMPANY_ADDRESS: string;
  SUPPLIER_NAME: string;
}